server/requirements.txt
openenv-core
fastapi
uvicorn[standard]
pydantic
websockets
openai